Change date values to prep for migrating to Eleventy
This way of writing dates in front matter is not valid in Eleventy, and
I don't really need time details in Jekyll anyway, so removing it now.
Hello World should be first, tweaking the 2nd post's date to make it so.

Also practicing writing regex is fun! Finally got it to find and
replace what it should and doing it manually would have been faster 🤣