Set workflow branch explicitly to gh-pages
Initially it seemed neat to skip branching and keep the settings for
GitHub Pages to automatically build from the main branch. But the
deploy commits then used main for the static output from the build —
and the repo is no longer an Eleventy project with version control of
files like eleventy.js and .github/workflows/build.yml 😅

Soooo now that I understand why there should be a branch for
publishing, I will explicitly set it to gh-pages instead of having the
branch be a hidden default which was just very confusing for me.