Fix all relative links within posts for Eleventy
When recently migrating from Jekyll, I added the pathPrefix for urls
get the  subdirectory from GitHub Pages ('notes') added to the link.
But that needs Eleventy url filter! Which is added in this commit.

Super pleased with figuring out this regex pattern 🦸🏻‍♀️
\s*{{\s*site.baseurl\s*}}(/*\d{4}/\d{2}/[a-z-_]+/)

to find my URLs formatted for Jekyll, and then using this for replace
{{ '$1' | url }}