Give all markdown tables header for new parser
Jekyll uses Kramdown, which is a parser that supports making tables
in markdown without a header — and so I made a bunch of those in
posts over the years. Turns out that most parsers don't support markdown
tables without a header. After migrating to Eleventy, these were no
longer parsed as tables. Easiest fix is to add these empty headers.